% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsAnalysis.R
\name{ATSC}
\alias{ATSC}
\title{Automated Trimmed & Sparse Clustering}
\usage{
ATSC(
  data,
  k.range = c(2, 15),
  bs = 100,
  cbi = "clara",
  max_alpha = 0.1,
  L1 = NULL,
  alpha = NULL,
  gold_standard = NULL,
  seed = NULL,
  numCores = 1,
  clusteringSparsity = "kmeans"
)
}
\arguments{
\item{data}{A \code{\link{SummarizedExperiment}}.
The SummarizedExperiment must contain an assay with the following structure:
A valid header with names. The first column of the header is the ID or name
of the instance of the dataset (e.g., ontology, pathway, etc.) on which the
metrics are measured.
The other columns of the header contains the names of the metrics.
The rows contains the measurements of the metrics for each instance in the dataset.}

\item{k.range}{Concatenation of two positive integers.
The first value \code{k.range[1]} is considered as the lower bound of the range,
whilst the second one, \code{k.range[2]}, as the higher. Both values must be
contained in [2,15] range.}

\item{bs}{Positive integer. Bootstrap value to perform the resampling.}

\item{cbi}{Clusterboot interface name (default: "clara"):
"kmeans", "clara", "clara_pam", "hclust", "pamk", "pamk_pam", "rskc".}

\item{max_alpha}{Maximum value of alpha, iterating over seq(0, max_alpha, 0.01).}

\item{L1}{A single L1 bound on weights (the feature weights), see \code{\link{RSKC}}.
If NULL, it is computed automatically.}

\item{alpha}{Trimming portion for RSKC, see \code{\link{RSKC}}.
If NULL, it is computed automatically via \code{max_alpha}.}

\item{gold_standard}{Numeric vector. A vector of clusters from a gold standard
classification. Only used when all metrics are clustered together.}

\item{seed}{Positive integer. A seed for internal bootstrap.}

\item{numCores}{Number of cores to be used for the alpha search (>1 will use parallel processing).}

\item{clusteringSparsity}{Clustering method used to determine the L1 bound (default: "kmeans").}
}
\value{
A list containing:
\item{stab}{A data frame containing standardized stability.}
\item{qual}{A data frame containing standardized quality.}
\item{optimalK}{The optimal k value representing the optimal number of clusters determined from the initial analysis.}
\item{stab_ATSC}{A data frame containing standardized stability after applying ATSC.}
\item{qual_ATSC}{A data frame containing standardized quality  applying ATSC.}
\item{optimalK_ATSC}{The optimal k value representing the optimal number of clusters determined after applying ATSC.}
\item{rskcOut}{An object returned by the RSKC function containing clustering results, including weights and trimmed observations.}
\item{trimmedRows}{A vector of indices representing the rows that were trimmed from the dataset during the clustering process.}
\item{trimmedColumns}{A vector of names representing the columns that were trimmed (i.e., removed) from the dataset due to zero weights.}
\item{trimmedDataset}{A data frame containing the final processed dataset after trimming rows and columns.}
}
\description{
Automated Trimmed & Sparse Clustering.
This methods performs an optimal k value analysis with \code{\link{stabilityRange}}, \code{\link{qualityRange}}
and \code{\link{getOptimalKValue}} evaluomeR methods.
The optimal \code{k} value is used to compute estimate a \code{L1} bound and an \code{alpha} trimming portion automatically
in order to perform an automatic trimmed and sparse clustering.
This posibily results in the input dataset being trimmed (either by columns, determined by \code{L1} or
by rows, determined by \code{alpha}).
Another optimal k value analysis is then executed over the trimmed dataset, to conclude with the an optimal partition.
}
