% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioMetrics}
\alias{bioMetrics}
\title{Dataset: Metrics for biological pathways}
\format{
An object of class \code{SummarizedExperiment} with 15 rows and 3 columns.
}
\usage{
data("bioMetrics")
}
\description{
Metrics for biological pathways, 2 metrics that
quantitative characterizations of the importance of regulation in biochemical
pathway systems, including systems designed for applications in synthetic
biology or metabolic engineering. The metrics are reachability and efficiency
}
\references{
\insertRef{davis2018metrics}{evaluomeR}
}
\keyword{datasets}
