% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{cleanDataset}
\alias{cleanDataset}
\title{Clean Dataset}
\usage{
cleanDataset(dataset, correlation_threshold = 1)
}
\arguments{
\item{dataset}{A data frame containing the dataset to be cleaned.}

\item{correlation_threshold}{A numeric value between -1 and 1 that specifies the correlation threshold
for removing perfectly correlated columns.}
}
\value{
A list containing the cleaned dataset (\code{dataset}) with the description column
and the correlation matrix R (\code{R}).
}
\description{
Cleans the dataset by removing non-numeric columns and those with perfect correlations.
}
