% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{evaluomeRSupportedCBI}
\alias{evaluomeRSupportedCBI}
\title{Get supported CBIs in evaluomeR.}
\usage{
evaluomeRSupportedCBI()
}
\value{
A String vector.
}
\description{
A vector of supported CBIs available in evaluomeR.
}
\examples{
supportedCBIs <- evaluomeRSupportedCBI

}
