% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsAnalysis.R
\name{getMetricRangeByCluster}
\alias{getMetricRangeByCluster}
\title{Get the range of each metric per cluster from the optimal cluster.
getMetricRangeByCluster}
\usage{
getMetricRangeByCluster(df, k.range, bs, seed)
}
\arguments{
\item{df}{Input data frame. The first column denotes the identifier of the
evaluated individuals. The remaining columns contain the metrics used to
evaluate the individuals. Rows with NA values will be ignored.}

\item{k.range}{Range of k values in which the optimal k will be searched}

\item{bs}{Bootstrap re-sample param.}

\item{seed}{Random seed to be used.}
}
\value{
A dataframe including the min and the max value for each
pair (metric, cluster).
}
\description{
Obtains the ranges of the metrics obtained by each optimal cluster.
}
