% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsAnalysis.R
\name{getRSKCAlpha}
\alias{getRSKCAlpha}
\title{Computes RSKC alpha
getRSKCAlpha}
\usage{
getRSKCAlpha(df, k, L1, seed = NULL)
}
\arguments{
\item{df}{Input data frame. The first column denotes the identifier of the
evaluated individuals. The remaining columns contain the metrics used to
evaluate the individuals. Rows with NA values will be ignored.}

\item{k}{K value (number of clusters)}

\item{L1}{A single L1 bound on weights (the feature weights), see \code{\link{RSKC}}.}

\item{seed}{Random seed to be used.}
}
\value{
Best suitable alpha.
}
\description{
Computes the proportion of the cases to be trimmed in robust sparse K-means, 0 <= alpha <= 1, see \code{\link{RSKC}}.
}
\examples{
data("ontMetrics")
alpha = getRSKCAlpha(ontMetrics, k=3, L1=2, seed=100)

}
