% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsAnalysis.R
\name{getRSKCL1Boundry}
\alias{getRSKCL1Boundry}
\title{Computes L1 boundry
getRSKCL1Boundry}
\usage{
getRSKCL1Boundry(df, k, seed = NULL)
}
\arguments{
\item{df}{Input data frame. The first column denotes the identifier of the
evaluated individuals. The remaining columns contain the metrics used to
evaluate the individuals. Rows with NA values will be ignored.}

\item{k}{K value (number of clusters)}

\item{seed}{Random seed to be used.}
}
\value{
A single L1 bound on weights (the feature weights), see \code{\link{RSKC}}.
}
\description{
Computes the L1 boundry for an RSKC cbi execution.
}
\examples{
data("ontMetrics")
l1_boundry = getRSKCL1Boundry(ontMetrics, k=3, seed=100)

}
