% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{golub}
\alias{golub}
\title{Dataset: Golub leukemia gene expression metrics}
\usage{
data("golub")
}
\description{
Bioinformatics metrics derived from the Golub et al. leukemia gene
expression dataset. Used for benchmarking ATSC and clustering methods.
}
\keyword{datasets}
