% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{loadSample}
\alias{loadSample}
\title{Sample input data loader}
\usage{
loadSample(descriptor)
}
\arguments{
\item{descriptor}{Sample file to load: "ont-metrics", "rna-metrics" or "biopathways-metrics".}
}
\value{
The dataset specified via \code{descriptor} as a dataframe.
}
\description{
This method is a wrapper to load sample input data located inside evaluomeR package.
}
\examples{
# Using example data from our package
metrics = loadSample("rna-metrics")

}
