% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsAnalysis.R
\name{metricsAnalysis}
\alias{metricsAnalysis}
\title{Global statistic analysis}
\usage{
metricsAnalysis(data)
}
\arguments{
\item{margins}{See \code{\link{par}}.}
}
\value{
A dataframe.
}
\description{
Global statistic analysis.
}
\examples{
# Using example data from our package
data("ontMetrics")
cor = metricsAnalysis(ontMetrics, getImages = TRUE, margins = c(1,0,5,11))

}
