% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nci60_k10}
\alias{nci60_k10}
\title{Dataset: NCI-60 cancer cell line metrics (k=10 partition)}
\usage{
data("nci60_k10")
}
\description{
Bioinformatics metrics computed on the NCI-60 cancer cell line panel,
pre-partitioned into 10 clusters. Used for benchmarking clustering methods.
}
\keyword{datasets}
