% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontMetrics}
\alias{ontMetrics}
\title{Dataset: Structural ontology metrics}
\format{An object of class \code{SummarizedExperiment} with 80 rows and 20 columns.}
\usage{
data("ontMetrics")
}
\description{
Structural ontology metrics, 19 metrics measuring
structural aspects of bio-ontologies have been analysed on two different
corpora of ontologies: OBO Foundry and AgroPortal
}
\references{
\insertRef{ontoeval}{evaluomeR}
}
\keyword{datasets}
