% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ontMetricsOBO}
\alias{ontMetricsOBO}
\title{Dataset: Structural ontology metrics (OBO Foundry subset)}
\usage{
data("ontMetricsOBO")
}
\description{
Structural ontology metrics measured on the OBO Foundry corpus of
bio-ontologies. A subset of the ontMetrics dataset restricted to OBO
Foundry ontologies.
}
\keyword{datasets}
