% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{PCASuitability}
\alias{PCASuitability}
\title{PCA Suitability}
\usage{
PCASuitability(R, sig_level = 0.05)
}
\arguments{
\item{R}{A correlation matrix of the dataset.}

\item{sig_level}{A numeric significance level for Bartlett's test, default is 0.05.}
}
\value{
A list indicating if PCA is suitable, along with the results of Bartlett's and KMO tests.
}
\description{
Performs Bartlett's test and KMO test to determine the suitability of PCA on the given correlation matrix R.
}
