% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{performPCA}
\alias{performPCA}
\title{Perform PCA}
\usage{
performPCA(dataset, ncp = 5, scale = TRUE, visualize = FALSE)
}
\arguments{
\item{dataset}{A data frame to be analyzed via PCA.}

\item{ncp}{An integer specifying the number of principal components to retain, default is 5.}

\item{scale}{A boolean indicating whether to scale the data, default is TRUE.}
}
\value{
A list containing the PCA results, summary of eigenvalues, contributions, coordinates, and more.
}
\description{
Executes PCA on the provided dataset and summarizes the results.
}
