% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{plotPCA_fviz_biplot}
\alias{plotPCA_fviz_biplot}
\title{PCA Biplot}
\usage{
plotPCA_fviz_biplot(pca_result, title = NULL)
}
\arguments{
\item{pca_result}{The result of a PCA analysis (object returned from `performPCA`).}

\item{title}{An optional title for the biplot. If NULL, a default title will be used.}
}
\value{
A ggplot object representing the PCA biplot.
}
\description{
Creates a biplot to visualize both the individuals and variables in a PCA result.
}
