% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{plotPCA_fviz_screeplot}
\alias{plotPCA_fviz_screeplot}
\title{Scree Plot for PCA}
\usage{
plotPCA_fviz_screeplot(pca_result, title = NULL)
}
\arguments{
\item{pca_result}{The result of a PCA analysis (object returned from `performPCA`).}

\item{title}{An optional title for the scree plot. If NULL, a default title will be used.}
}
\value{
A ggplot object representing the scree plot.
}
\description{
Creates a scree plot to visualize the explained variance by principal components from a PCA result.
}
