% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityIndices.R
\name{quality}
\alias{quality}
\title{Goodness of classifications.}
\usage{
quality(data, k = 5, getImages = TRUE, label = NULL, path = NULL)
}
\arguments{
\item{data}{A matrix. The first row is the header. The first
column of the header is the ID or name of the instance of the dataset
(e.g., ontology, pathway, etc.) on which the metrics are measured.
The other columns of the header contains the names of the metrics.
The rows contains the measurements of the metrics for each instance in the dataset.}

\item{k}{Positive integer. Number of clusters between [2,15] range.}

\item{getImages}{Boolean. If true, a plot is displayed.}

\item{label}{String. If not NULL, the label will appear on the title of the plots.}

\item{path}{String. Path to a valid directory where plots are saved.}
}
\value{
A dataframe containing the silhouette width measurements and
cluster sizes for cluster \code{k}.
}
\description{
The goodness of the classifications are assessed by validating the clusters
generated. For this purpose, we use the Silhouette width as validity index.
This index computes and compares the quality of the clustering outputs found
by the different metrics, thus enabling to measure the goodness of the
classification for both instances and metrics. More precisely, this goodness measurement
provides an assessment of how similar an instance is to other instances from
the same cluster and dissimilar to all the other clusters. The average on all
the instances quantifies how appropriately the instances are clustered. Kaufman
and Rousseeuw suggested the interpretation of the global Silhouette width score
as the effectiveness of the clustering structure. The values are in the
range [0,1], having the following meaning:

\itemize{
\item There is no substantial clustering structure: [-1, 0.25].
\item The clustering structure is weak and could be artificial: ]0.25, 0.50].
\item There is a reasonable clustering structure: ]0.50, 0.70].
\item A strong clustering structure has been found: ]0.70, 1].
}
}
\examples{
# Using example data from our package
metrics = loadSample("ont-metrics")
result = quality(data=metrics, k=5)
result = quality(metrics, k=6, getImages=FALSE)
result = quality(metrics, k=4, label="Experiment 1:")

}
\references{
\insertRef{kaufman2009finding}{evaluomeR}
}
