% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{removeCorrelations}
\alias{removeCorrelations}
\title{Remove Correlations}
\usage{
removeCorrelations(numeric_dataset, correlation_threshold = 1)
}
\arguments{
\item{numeric_dataset}{A data frame consisting of only numeric columns.}

\item{correlation_threshold}{A numeric threshold for identifying perfect correlations.}
}
\value{
A list containing the cleaned dataset and the remaining column names, along with the correlation matrix R.
}
\description{
Identifies and removes  correlated columns from the numeric dataset, according to `correlation_threshold`.
}
