% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreprocessing.R
\name{removeNonNumericColumns}
\alias{removeNonNumericColumns}
\title{Remove Non-Numeric Columns}
\usage{
removeNonNumericColumns(dataset)
}
\arguments{
\item{dataset}{A data frame from which non-numeric columns are to be removed.}
}
\value{
A data frame containing only numeric columns. Also prints non-numeric columns found.
}
\description{
This function removes non-numeric columns from the dataset and returns only numeric columns.
}
