% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rnaMetrics}
\alias{rnaMetrics}
\title{Dataset: RNA quality metrics}
\format{
An object of class \code{SummarizedExperiment} with 16 rows and 3 columns.
}
\usage{
data("rnaMetrics")
}
\description{
RNA quality metrics for the assessment of gene expression
differences, 2 quality metrics from 16 aliquots of a unique batch of RNA
Samples. The metrics are Degradation Factor (DegFact) and RNA Integrity Number
(RIN)
}
\references{
\insertRef{imbeaud2005towards}{evaluomeR}
}
\keyword{datasets}
