% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{seToDataFrame}
\alias{seToDataFrame}
\title{SummarizedExperiment to Dataframe}
\usage{
seToDataFrame(SummarizedExperiment)
}
\arguments{
\item{SummarizedExperiment}{A \code{SummarizedExperiment} object
(see \code{\link{SummarizedExperiment}}).}
}
\value{
The dataframe that contains information of the first
assay in \code{SummarizedExperiment}.
}
\description{
This method is a wrapper to transform a SummarizedExperiment object to a
Dataframe processable in our methods.
}
\examples{
# Using example data from airway package
library(airway)
data(airway)
airwayData = seToDataFrame(airway)
airwayData = airwayData[1:10000,1:4]
stability(airwayData, bs = 20, getImages=FALSE)
correlations(airwayData, getImages=FALSE)

}
